refactor: password item card
